Added Java Tutorial Course
As a resource, I've included two Java tutorial courses from Scaler Topics - one in video format and the other in text. These courses are intended for beginners who want to begin their Java learning journey, and they come with a free certification.

The course in text format is composed of 12 modules:

1. Introduction to Java
2. Getting Started with the Syntax
3. Variables and methods in Java
4. Operators in Java
5. Conditional Flow of Control in Java
6. Arrays in Java
7. String Class in Java
8. Collection Framework in Java
9. OOPs Concepts in Java
10. Exception Handling in Java
11. Java IO Stream
12. Lambda and Streams in Java

Hope this helps the learner.