Added Java execution process to java fundamentals section.
This PR adds a section to the Java Fundamentals section of the roadmap.sh project that outlines the necessary steps to work with Java programming on a computer. With this addition, developers can quickly catch up on the fundamentals of Java programming and gain a better understanding of how to work with Java on their machines. Additionally, a clean and understandable graphic is included to help visualize the process.