Update set.md
The statement, "It is a collection of elements where each element is stored as a value without any keys," is slightly ambiguous. It's important to clarify that while Set does not use keys in the way that an `Object` or `Map` does, each element in a Set can be thought of as both a key and a value, which is distinct from how keys and values are defined in objects or maps.