Added roadmap-list.cjs for compiling textual roadmaps
Helper script to generate a structured Markdown output from any roadmap. This can be useful for keeping a textual version of a roadmap in personal notes and ticking off resources as learning progresses.

It has two modes - 'simple' (default) and 'full', could potentially merged with `roadmap-links.cjs` in the future.