Propose to add content to 103-docker-and-oci.md
The OCI-specification has three core specifications while on the description written only two are mentioned. 

I added the *Distribution Specification (distribution-spec)*.

According to the website, the distro-spec defines a standardized API protocol for distributing container images and other content. It is primarly designed to ensure interoperability and consistency across different container registries and tools, regardless of their specific implementation.

I attach below links to the website of OCI and the github repo
- [https://github.com/opencontainers/distribution-spec](On the github repo of the OCI) 
- [https://opencontainers.org/about/overview/](Website of the OCI)