change the CEH position to CISSP
This is a suggestion, I believe that CEH appears as a highlight due to the larger space and this bothers many people.

CISSP is a top professional-level certification in the field of cybersecurity and the the most respected certificate in the IT security field. He really deserves his place in the spotlight.