ADD PythonDataScienceHandbook For data-analyst
Python Data Science Handbook is a well-known book about Python data processing and analysis. Its original author has made the contents of this book and the corresponding notebook source code open source on GitHub. It is worth a look, especially for those who want to understand or urgently need Python's functions in data processing and analysis.