Add a book (#6756)
* ADD PythonDataScienceHandbook For data-analyst

Python Data Science Handbook is a well-known book about Python data processing and analysis. Its original author has made the contents of this book and the corresponding notebook source code open source on GitHub. It is worth a look, especially for those who want to understand or urgently need Python's functions in data processing and analysis.

* Update src/data/roadmaps/data-analyst/content/learn-a-programming-lang@i2uEcaO4bJhcZ5ayRs2CQ.md

---------

Co-authored-by: Kamran Ahmed <kamranahmed.se@gmail.com>