fix(cyber-security): clarify zero-day vulnerability description
* Update zero-day@v9njgIxZyabJZ5iND3JGc.md

a software security flaw unknown to the vendor and its developers, not exploit developers. Exploit developers maybe already known about it, but they are not part of the development team in charge of the application.

* Update zero-day@zqRaMmqcLfx400kJ-h0LO.md

a software security flaw unknown to the vendor and its developers, not exploit developers. Exploit developers maybe already known about it, but they are not part of the development team in charge of the application.